refactor: 增加hook useCollapseTransition
