+++ packages/varlet-ui/src/collapse-transition/CollapseTransition.vue
@@ -1,16 +1,22 @@
 <template>
   <div :class="n()">
-    <div :class="n('content')" v-show="showContent" ref="contentEl" @transitionend="handleTransitionend">
+    <div
+      :class="n('content')"
+      v-show="showContent"
+      ref="contentEl"
+      @transitionend="handleTransitionEnd"
+      @transitionstart="handleTransitionStart"
+    >
       <slot />
     </div>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, ref, watch, nextTick } from 'vue'
-import { raf } from '@varlet/shared'
+import { defineComponent, ref, watch, nextTick, computed } from 'vue'
 import { createNamespace } from '../utils/components'
 import { props } from './props'
+import { useCollapseTransition } from './useCollapseTransition'
 
 const { name, n, classes } = createNamespace('collapse-transition')
 
@@ -20,9 +26,10 @@ export default defineComponent({
   setup(props) {
     const showContent = ref(false)
     const contentEl = ref<HTMLDivElement | null>(null)
+    const expand = computed(() => props.expand)
 
     watch(
-      () => props.expand,
+      expand,
       (value) => {
         nextTick().then(() => {
           value ? openPanel() : closePanel()
@@ -31,55 +38,19 @@ export default defineComponent({
       { immediate: true }
     )
 
-    async function openPanel() {
-      if (!contentEl.value) {
-        return
-      }
-
-      contentEl.value.style.height = ''
-      showContent.value = true
-      await raf()
-
-      if (!contentEl.value) {
-        return
-      }
-
-      const { offsetHeight } = contentEl.value
-      contentEl.value.style.height = 0 + 'px'
-      await raf()
-
-      if (!contentEl.value) {
-        return
-      }
-
-      contentEl.value.style.height = offsetHeight + 'px'
-    }
-
-    async function closePanel() {
-      if (!contentEl.value) {
-        return
-      }
-
-      const { offsetHeight } = contentEl.value
-      contentEl.value.style.height = offsetHeight + 'px'
-      await raf()
-      contentEl.value.style.height = 0 + 'px'
-    }
-
-    function handleTransitionend() {
-      if (!props.expand) {
-        showContent.value = false
-      }
-
-      contentEl.value!.style.height = ''
-    }
+    const { openPanel, closePanel, handleTransitionEnd, handleTransitionStart } = useCollapseTransition(
+      contentEl,
+      showContent,
+      expand
+    )
 
     return {
       showContent,
       contentEl,
       n,
       classes,
-      handleTransitionend,
+      handleTransitionEnd,
+      handleTransitionStart,
     }
   },
 })
